+++ expenses/src/sdk/sdk.ts
@@ -50,7 +50,7 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "prealpha";
-    sdkVersion = "0.31.0";
+    sdkVersion = "0.31.1";
     genVersion = "2.55.0";
 
     public constructor(init?: Partial<SDKConfiguration>) {