+++ expenses/RELEASES.md
@@ -567,3 +567,11 @@ Based on:
 - Speakeasy CLI 1.52.0 (2.55.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v0.31.0] https://www.npmjs.com/package/@codat/sync-for-expenses/v/0.31.0 - expenses
+
+## 2023-07-04 00:25:33
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Expenses.yaml
+- Speakeasy CLI 1.52.0 (2.55.0) https://github.com/speakeasy-api/speakeasy
+### Releases
+- [NPM v0.31.1] https://www.npmjs.com/package/@codat/sync-for-expenses/v/0.31.1 - expenses